requirements.txt
# Core data fetching
yfinance>=0.2.18
pandas>=2.0.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Sentiment analysis
newsapi-python>=0.2.6
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: ML and analysis (for Jupyter container)
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter for ML development
jupyterlab>=4.0.0
notebook>=7.0.0

# Optional: Database connections
psycopg2-binary>=2.9.0
redis>=4.5.0

# Optional: API rate limiting and caching
ratelimit>=2.2.0